requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
gunicorn==21.2.0